# .hacktogether/runs/87edeba4-9a80-4a73-858b-548fd9026da4/03-api-design/openapi.yaml
openapi: 3.1.0
info:
  title: Harvoost API
  version: 1.0.0
  description: |
    Harvoost — Harvest-style time-tracking SaaS for a single tenant on Azure.
    Provides endpoints for time entries, mood capture, scheduling, leave booking,
    two-stage approvals, exception handling, financial dashboards, Harvest-compatible
    Excel export, an RBAC-aware manager chatbot, and autonomous weekly summaries.

    ## Authentication
    Bearer token (opaque server-issued session id) in the `Authorization` header.
    The token is minted by `apps/api` after the Entra ID OIDC handshake completes.
    Token lifetime is 12h; clients refresh via `POST /v1/auth/refresh` (separate spec).

    ## Field casing
    All JSON field names, path parameters, and query parameters use `snake_case`.

    ## Error envelope
    Uniform shape across the entire API: `{ code, message, details? }`. See `ErrorResponse` schema.

    ## Pagination
    - **Cursor-based** for high-volume / append-mostly resources: time_entries, audit_log,
      exceptions, chatbot_conversations, chatbot_messages, leave_requests.
    - **Offset-based** for catalog-style resources: users, projects, clients, cost_rates, billable_rates.

    ## Idempotency
    `Idempotency-Key` header is REQUIRED on `POST /v1/time-entries/start`, `/stop`, and
    `/switch`. The server stores `(user_id, idempotency_key) → response` for 5 minutes;
    retried requests with the same key return the original response. Reuse of the same
...

    ## ISO 8601
...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/cost-rates:
    get:
      tags: [Rates]
      summary: List employee cost rates (Admin/FinMgr only)
      description: |
        Effective-dated per-user cost rates from `employee_cost_rates`. Admin/FinMgr ONLY — cost
        is financial data; Manager/Employee receive 403 `RBAC_FORBIDDEN`. The web Rates page is
        `canSeeFinancialData`-gated to mirror this server-side gate.

        Two usage modes (both offset-paginated):
        - `current=true` → the single active rate per user (`effective_from <= today AND
          (effective_to IS NULL OR effective_to > today)`).
        - `user_id=<id>` → that user's full rate history (newest first).
      operationId: listCostRates
      parameters:
        - in: query
          name: current
          schema: { type: boolean }
          description: When true, returns only the currently-effective rate per user.
        - in: query
          name: user_id
          schema: { type: integer, format: int64 }
          description: Restrict to one user's full rate history.
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Offset-paginated cost rates.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/OffsetPaginationMeta'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/CostRate' }
                    required: [data]
              example:
                page: 1
                page_size: 50
                total_count: 12
                data:
                  - id: 301
                    user_id: 18
                    user_display_name: Bob Example
                    rate: 750
                    currency: ZAR
                    effective_from: 2026-01-01
                    effective_to: null
                    created_by: 1
                    created_at: 2026-01-01T08:00:00+00:00
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Rates]
      summary: Set a new effective-dated cost rate (Admin/FinMgr only)
      description: |
        Creates a new effective-dated cost rate for a user. The server end-dates the prior open
        row and inserts the new one; the `ecr_no_overlap` GiST exclusion enforces that effective
        ranges never overlap. `created_by` is set from the actor; an audit entry is written.

        On an overlapping effective range (Postgres `23P01` exclusion violation), the server maps
        the error to a clean 409 / 422 `VALIDATION_FAILED` rather than leaking the DB error.
      operationId: createCostRate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCostRateRequest' }
            example:
              user_id: 18
              rate: 800
              currency: ZAR
              effective_from: 2026-06-01
      responses:
        '201':
          description: Cost rate created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CostRate' }
              example:
                id: 312
                user_id: 18
                user_display_name: Bob Example
                rate: 800
                currency: ZAR
                effective_from: 2026-06-01
                effective_to: null
                created_by: 1
                created_at: 2026-05-23T09:00:00+00:00
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Overlapping effective range for this user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
              example:
                code: VALIDATION_FAILED
                message: A cost rate already covers this effective range for user 18.
        '422': { $ref: '#/components/responses/ValidationFailed' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Opaque server-issued session token (the API stores sessions in the `sessions` table
        — the token itself does not encode claims). The `bearerFormat: JWT` annotation here is
        a convention for OpenAPI tooling; the actual format is opaque (32+ bytes base64url).
        Issued by `POST /v1/auth/oidc/callback` after a successful Entra ID OIDC handshake.
  parameters:
...
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 1-based page index for offset-paginated collections.
    PageSize:
      in: query
      name: page_size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Items per page (1..200, default 50).
...
  responses:
    BadRequest:
      description: Malformed request (e.g., bad JSON, bad date format).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: VALIDATION_FAILED
            message: Field `start_at` is not a valid ISO 8601 datetime with offset.
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: VALIDATION_FAILED
            message: Missing Authorization header.
    Forbidden:
      description: |
        Authenticated but not allowed (RBAC). Out-of-scope drill-downs return this code.
        Note: chatbot "no data accessible" responses are 200s with a polite text reply,
        NOT 403s.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: RBAC_FORBIDDEN
            message: You do not have access to this resource.
...
    ValidationFailed:
      description: Request body or query parameters failed validation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            code: VALIDATION_FAILED
            message: One or more fields failed validation.
            details:
              fields:
                - field: end_date
                  error: must be >= start_date
...
  schemas:
...
    ErrorCode:
      type: string
      enum:
        - RBAC_FORBIDDEN
        - ENTRY_LOCKED
        - CHATBOT_DISABLED
        - IDEMPOTENCY_CONFLICT
        - VALIDATION_FAILED
        - NOT_FOUND
        - RATE_LIMITED
        - LLM_UNAVAILABLE
        - OIDC_FAILURE
        - K_ANONYMITY_THRESHOLD
      description: |
        Canonical machine-readable error codes returned in the `code` field of every error
        envelope. See API_NOTES.md for the semantic meaning of each code.

    # ---------- Error & pagination ----------
    ErrorResponse:
      type: object
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message:
          type: string
          description: Human-readable English explanation. Safe to display to end users.
        details:
          description: Optional structured context (e.g., `fields[]` for validation errors).
          oneOf:
            - { type: object, additionalProperties: true }
            - { type: array, items: { type: object, additionalProperties: true } }
      required: [code, message]
      examples:
        - code: ENTRY_LOCKED
          message: Cannot edit entry — currently in status manager_approved.
        - code: RBAC_FORBIDDEN
          message: You do not have access to this resource.
...
    OffsetPaginationMeta:
      type: object
      description: Offset-based pagination metadata.
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total_count: { type: integer }
      required: [page, page_size, total_count]

    # ---------- Auth ----------
...
    CostRate:
      type: object
      description: |
        Effective-dated per-user cost rate (`employee_cost_rates`). Financial data — only
        returned to Admin/FinMgr.
      properties:
        id: { type: integer, format: int64 }
        user_id: { type: integer, format: int64 }
        user_display_name:
          oneOf:
            - { type: string }
            - { type: 'null' }
          description: Convenience join for the admin table; may be absent on POST echoes.
        rate: { type: number, format: float }
        currency: { type: string, minLength: 3, maxLength: 3, description: ISO 4217. }
        effective_from: { type: string, format: date }
        effective_to:
          oneOf:
            - { type: string, format: date }
            - { type: 'null' }
          description: '`null` for the currently-open (active) row.'
        created_by: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
      required: [user_id, rate, currency, effective_from, created_by, created_at]
    CreateCostRateRequest:
      type: object
      properties:
        user_id: { type: integer, format: int64 }
        rate: { type: number, format: float, minimum: 0 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        effective_from: { type: string, format: date }
      required: [user_id, rate, currency, effective_from]
...